Cleanup [Jobs panel] remove redundant components
