--- src/components/ScheduleJob/scheduleJob.scss
@@ -1,7 +1,0 @@
-@import '~igz-controls/scss/colors';
-@import '~igz-controls/scss/borders';
-@import '~igz-controls/scss/mixins';
-
-.jobs-panel__schedule {
-  @include schedule;
-}
