--- src/components/ScheduleJobSimple/ScheduleJobSimple.js
@@ -1,80 +1,0 @@
-/*
-Copyright 2019 Iguazio Systems Ltd.
-
-Licensed under the Apache License, Version 2.0 (the "License") with
-an addition restriction as set forth herein. You may not use this
-file except in compliance with the License. You may obtain a copy of
-the License at http://www.apache.org/licenses/LICENSE-2.0.
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
-implied. See the License for the specific language governing
-permissions and limitations under the License.
-
-In addition, you may not use the software for any purposes that are
-illegal under applicable law, and the grant of the foregoing license
-under the Apache 2.0 license is conditioned upon your compliance with
-such restriction.
-*/
-import React from 'react'
-import PropTypes from 'prop-types'
-
-import CheckBox from '../../common/CheckBox/CheckBox'
-import ScheduleRecurring from '../../elements/ScheduleRecurring/ScheduleRecurring'
-import DatePicker from '../../common/DatePicker/DatePicker'
-import TimePicker from '../../common/TimePicker/TimePicker'
-
-const ScheduleJobSimple = ({
-  date,
-  daysOfWeek,
-  handleDaysOfWeek,
-  isRecurring,
-  recurringDispatch,
-  recurringState,
-  setDate,
-  setIsRecurring,
-  setTime,
-  time
-}) => {
-  return (
-    <>
-      <div className="input-container">
-        <DatePicker onChange={setDate} date={date} />
-        <TimePicker onChange={setTime} value={time} />
-      </div>
-      <div className="checkbox-container">
-        <CheckBox
-          item={{ label: 'Recurring', id: 'recurring' }}
-          onChange={recurring =>
-            setIsRecurring(state => (state === recurring ? '' : recurring))
-          }
-          selectedId={isRecurring}
-        />
-      </div>
-      {isRecurring && (
-        <ScheduleRecurring
-          daysOfWeek={daysOfWeek}
-          handleDaysOfWeek={handleDaysOfWeek}
-          recurringDispatch={recurringDispatch}
-          recurringState={recurringState}
-        />
-      )}
-    </>
-  )
-}
-
-ScheduleJobSimple.propTypes = {
-  date: PropTypes.oneOfType([PropTypes.string, PropTypes.instanceOf(Date)]),
-  daysOfWeek: PropTypes.array.isRequired,
-  handleDaysOfWeek: PropTypes.func.isRequired,
-  isRecurring: PropTypes.string.isRequired,
-  recurringDispatch: PropTypes.func.isRequired,
-  recurringState: PropTypes.shape({}).isRequired,
-  setDate: PropTypes.func.isRequired,
-  setIsRecurring: PropTypes.func.isRequired,
-  setTime: PropTypes.func.isRequired,
-  time: PropTypes.string.isRequired
-}
-
-export default ScheduleJobSimple
