+++ src/components/SheduleWizard/ScheduleWizard.js
@@ -26,12 +26,11 @@ import ScheduleWizardSimple from './ScheduleWizardSimple'
 import ScheduleWizardCronstring from './ScheduleWizardCronstring'
 import ErrorMessage from '../../common/ErrorMessage/ErrorMessage'
 
-import { tabs } from '../ScheduleJob/scheduleJobData'
 import { decodeLocale, getWeekDays, getWeekStart } from '../../utils/datePicker.util'
 import { getFormatTime } from '../../utils'
 import { generateCronInitialValue } from '../../utils/generateCronInitialValue'
 import { PRIMARY_BUTTON, TERTIARY_BUTTON } from 'igz-controls/constants'
-import { getDefaultSchedule, scheduleDataInitialState } from './scheduleWizard.util'
+import { getDefaultSchedule, scheduleDataInitialState, tabs } from './scheduleWizard.util'
 import { SCHEDULE_DATA } from '../../types'
 import { SIMPLE_SCHEDULE, CRONSTRING_SCHEDULE } from '../../constants'
