--- src/elements/JobsPanelCredentialsAccessKey/jobsPanelCredentialsAccessKey.scss
@@ -1,8 +1,0 @@
-.job-panel__item.access-key {
-  height: 64px;
-  padding: 22px 40px;
-
-  &.without-padding {
-    padding: 0;
-  }
-}
