+++ src/utils/getDefaultSchedule.js
@@ -17,7 +17,7 @@ illegal under applicable law, and the grant of the foregoing license
 under the Apache 2.0 license is conditioned upon your compliance with
 such restriction.
 */
-import { scheduleActionType } from '../components/ScheduleJob/recurringReducer'
+import { scheduleActionType } from '../components/FeatureSetsPanel/ScheduleFeatureSet/recurringReducer'
 
 export const getDefaultSchedule = (defaultCron, recurringDispatch) => {
   let cron = defaultCron.split(' ')
